LPD-19790 - Refactor. Extract shared types & constants
Replace references
