+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/cards/Cards.tsx
@@ -19,9 +19,7 @@ import {API_URL, OBJECT_RELATIONSHIP} from '../../../utils/constants';
 import openDefaultFailureToast from '../../../utils/openDefaultFailureToast';
 import openDefaultSuccessToast from '../../../utils/openDefaultSuccessToast';
 import {IField} from '../../../utils/types';
-import {IBaseVisualizationMode} from '../VisualizationModes';
 
-export interface ICards extends IBaseVisualizationMode<'cards'> {}
 interface IFDSCardsSection {
 	externalReferenceCode: string;
 	fieldName: string;
