+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/table/Table.tsx
@@ -40,12 +40,10 @@ import '../../../../css/TableVisualizationMode.scss';
 import ClayIcon from '@clayui/icon';
 
 import {IFDSField, IField} from '../../../utils/types';
-import {IBaseVisualizationMode} from '../VisualizationModes';
 import AddFieldsModalContent from './modal_content/AddFieldsModalContent';
 
 const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
-export interface ITable extends IBaseVisualizationMode<'table'> {}
 interface ISaveFDSFieldsModalContentProps {
 	closeModal: Function;
 	fdsFields: Array<IFDSField>;
