+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/VisualizationModes.d.ts
@@ -6,11 +6,6 @@
 /// <reference types="react" />
 
 import {IFDSViewSectionProps} from '../../FDSView';
-export interface IBaseVisualizationMode<Type extends string> {
-	label: string;
-	type: Type;
-	visualizationModeId: string;
-}
 export default function VisualizationModes(
 	props: IFDSViewSectionProps
 ): JSX.Element;
