+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/cards/Cards.d.ts
@@ -7,6 +7,4 @@
 
 import '../../../../css/CardsVisualizationMode.scss';
 import {IFDSViewSectionProps} from '../../../FDSView';
-import {IBaseVisualizationMode} from '../VisualizationModes';
-export interface ICards extends IBaseVisualizationMode<'cards'> {}
 export default function Cards(props: IFDSViewSectionProps): JSX.Element;
