+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/list/List.d.ts
@@ -7,6 +7,4 @@
 
 import '../../../../css/ListVisualizationMode.scss';
 import {IFDSViewSectionProps} from '../../../FDSView';
-import {IBaseVisualizationMode} from '../VisualizationModes';
-export interface IList extends IBaseVisualizationMode<'list'> {}
 export default function List(props: IFDSViewSectionProps): JSX.Element;
