+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/table/Table.d.ts
@@ -7,8 +7,6 @@
 
 import {IFDSViewSectionProps} from '../../../FDSView';
 import '../../../../css/TableVisualizationMode.scss';
-import {IBaseVisualizationMode} from '../VisualizationModes';
-export interface ITable extends IBaseVisualizationMode<'table'> {}
 declare function Table({
 	fdsClientExtensionCellRenderers,
 	fdsView,
