+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/utils/types.d.ts
@@ -20,6 +20,16 @@ export declare enum EFieldType {
 	OBJECT = 'object',
 	STRING = 'string',
 }
+export interface IBaseVisualizationMode<Type extends string> {
+	label: string;
+	thumbnail: string;
+	type: Type;
+	visualizationModeId: string;
+}
+export interface ICards extends IBaseVisualizationMode<'cards'> {}
+export interface IList extends IBaseVisualizationMode<'list'> {}
+export interface ITable extends IBaseVisualizationMode<'table'> {}
+export declare type TVisualizationMode = ICards | IList | ITable;
 export interface IField {
 	children?: Array<IField>;
 	format?: EFieldFormat;
